added (as yet untested) functionality to optionally log community source of a message and also not to print AuxSrc (even if requested) if it is empty.
MarkExternalCommunityMessages is now a configuration block parameter

This is in response to mikerb's request below:

When MarkExternalCommunityMessages =true, the logger will log the community name ONLY if the community name is different than the community of the MOOSDB.

By example, the following:

58.562          DEPLOY    pFooBar            PARK
becomes:
58.562          DEPLOY    pFooBar@shoreside  PARK

ONLY if "shoreside" is an external community, presumably from pMOOSBridge etc.  If the community is local, there is NO change to the log format for that line. If there happens to be a non-null auxilliary source as well, it would be "pFooBar:auxsrc@shoreside".

Signed-off-by: Paul Newman <pnewman@robots.ox.ac.uk>